+++ unnamed_engine/src/lib.rs
@@ -5,5 +5,6 @@ pub mod gui;
 pub mod input;
 pub mod networking;
 pub mod renderer;
+pub mod voxel;
 
 pub use log;